requirements.txt
pillow
pyzbar
cbor2
